+++ src/main/java/insane96mcp/iguanatweaksreborn/mixin/DigDurabilityEnchantmentMixin.java
@@ -11,11 +11,10 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
 @Mixin(DigDurabilityEnchantment.class)
 public abstract class DigDurabilityEnchantmentMixin {
-	@Inject(at = @At(value = "RETURN", ordinal = 1), method = "shouldIgnoreDurabilityDrop", cancellable = true)
+	@Inject(at = @At(value = "RETURN"), method = "shouldIgnoreDurabilityDrop", cancellable = true)
 	private static void onChanceToIgnoreDurabilityDrop(ItemStack stack, int lvl, RandomSource random, CallbackInfoReturnable<Boolean> cir) {
-		if (EnchantmentsFeature.isUnbreakingOverhaul()) {
-			cir.setReturnValue(random.nextFloat() < EnchantmentsFeature.unbreakingBonus(lvl));
-		}
+		if (EnchantmentsFeature.isUnbreakingOverhaul())
+			cir.setReturnValue(false);
 	}
 
 	@Inject(method = "getMaxLevel", at = @At("RETURN"), cancellable = true)