+++ src/main/java/insane96mcp/iguanatweaksreborn/module/experience/enchantments/enchantment/Luck.java
@@ -2,13 +2,13 @@ package insane96mcp.iguanatweaksreborn.module.experience.enchantments.enchantmen
 
 import insane96mcp.iguanatweaksreborn.data.generator.ISOItemTagsProvider;
 import insane96mcp.iguanatweaksreborn.module.experience.enchantments.EnchantmentsFeature;
-import insane96mcp.insanelib.base.Feature;
 import net.minecraft.tags.TagKey;
 import net.minecraft.world.entity.EquipmentSlot;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.enchantment.EnchantmentCategory;
 import net.minecraft.world.item.enchantment.LootBonusEnchantment;
+import org.jetbrains.annotations.NotNull;
 
 import java.util.Set;
 
@@ -16,17 +16,32 @@ public class Luck extends LootBonusEnchantment {
     public static final TagKey<Item> ACCEPTS_ENCHANTMENT = ISOItemTagsProvider.create("enchanting/accepts_luck");
     static final EnchantmentCategory CATEGORY = EnchantmentCategory.create("luck", item -> item.builtInRegistryHolder().is(ACCEPTS_ENCHANTMENT));
     public Luck() {
-        super(Rarity.RARE, CATEGORY, EquipmentSlot.MAINHAND);
+        super(Rarity.VERY_RARE, CATEGORY, EquipmentSlot.MAINHAND);
     }
 
     @Override
-    public boolean canApplyAtEnchantingTable(ItemStack stack) {
-        return super.canApplyAtEnchantingTable(stack) && EnchantmentsFeature.replaceBonusLootEnchantments && Feature.isEnabled(EnchantmentsFeature.class);
+    public int getMaxLevel() {
+        return 1;
+    }
+
+    @Override
+    public int getMinCost(int pEnchantmentLevel) {
+        return super.getMinCost(2);
+    }
+
+    @Override
+    public int getMaxCost(int pEnchantmentLevel) {
+        return super.getMaxCost(2);
+    }
+
+    @Override
+    public boolean canApplyAtEnchantingTable(@NotNull ItemStack stack) {
+        return super.canApplyAtEnchantingTable(stack) && EnchantmentsFeature.isBonusLootEnchantmentReworkEnabled();
     }
 
     @Override
     public boolean isDiscoverable() {
-        return super.isDiscoverable() && EnchantmentsFeature.replaceBonusLootEnchantments && Feature.isEnabled(EnchantmentsFeature.class);
+        return super.isDiscoverable() && EnchantmentsFeature.isBonusLootEnchantmentReworkEnabled();
     }
 
     @Override